requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
playwright==1.40.0
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.25.1
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
tenacity==8.2.3
pyotp==2.9.0
google-generativeai==0.3.2
apscheduler==3.10.4
